Agent & Obstacle thickness
Added more granularity to Agent radiuses & implemented obstacle thickness handling.
Agent now have two distinct radiuses, one used for agent-agent collision (radius), and another for agent-obstacle collision. This allow to control the "distance" at which an agent start colliding with an obstacle, on an individual basis.

Obstacles now have a thickness of their own, accounted for in the simulation.